% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-density_estimation.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{Compute quantiles for specified 'variable' grouping by 'ids' columns
 in 'variable'. The quantile for a real numer x is the fraction of
 values for 'variable' that are less than x.}
\usage{
compute_quantiles(data, ids, variable, probs = ppoints(1000),
  min_count = 15)
}
\description{
If lengths(probs)==1, then "Ordinates for Probability Plotting" with
 'probs' probability points are used.
}
\details{
Returns data.frame with the following columns
   ids, probs, quantiles, counts

# for example:
 q <- compute_quantiles(f, c("id_col1", "id_col2"), "var")

ggplot(q) + geom_line(aes(x=quantiles*180/pi, y=probs))
}

