% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-comparison_statistics.R
\name{cross_validate_statistics}
\alias{cross_validate_statistics}
\title{Evaluate a two sample tests between different classes of samples
conditional on distinct groups of identifying variables.
  The class.vars is used to identify the classes of the samples
  The id.vars is used to split the classes in to comparison groups}
\usage{
cross_validate_statistics(f, cv.var, id.vars, measure.vars, comp_fun)
}
\description{
Evaluate a two sample tests between different classes of samples
conditional on distinct groups of identifying variables.
  The class.vars is used to identify the classes of the samples
  The id.vars is used to split the classes in to comparison groups
}

