% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-primary_modes.R
\name{find_modes_in_density}
\alias{find_modes_in_density}
\title{provides estimate_primary_mode_1d, which estimates the location of
	the primary mode by binary searching the bandwidth of kernel
	density estimation till there is just one mode.}
\usage{
find_modes_in_density(dens_y)
}
\description{
provides estimate_primary_mode_1d, which estimates the location of
	the primary mode by binary searching the bandwidth of kernel
	density estimation till there is just one mode.
}

