% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-comparison_statistics.R
\name{kolmogorov_smirnov_test}
\alias{kolmogorov_smirnov_test}
\title{Prefer to use the anderson_darling_2_sample comparison
cf https://asaip.psu.edu/Articles/beware-the-kolmogorov-smirnov-test}
\usage{
kolmogorov_smirnov_test(a, b)
}
\description{
Prefer to use the anderson_darling_2_sample comparison
cf https://asaip.psu.edu/Articles/beware-the-kolmogorov-smirnov-test
}

