% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-coordinate_normalizations.R
\name{major_long_coords}
\alias{major_long_coords}
\title{These the coordinates for the Lambert-Azmuthal plots:
  longitude (around) 30, 60, 90, 120 degrees
  latitude  (in-out) 30, 60, 90, 120 degrees (starting from the positive x-axis)}
\usage{
major_long_coords(lat = c(pi/6, pi/2))
}
\description{
Equal Area Coordinate Grids
}

