% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-instancer.R
\name{prepare_feature_instances}
\alias{prepare_feature_instances}
\title{prepare instances to be viewed in PyMol script to look at feature instances
 The residues_of_interest should be a data.frame with the following columns:
   tag, id, chain, resNum, atom
 See the error message below for more details}
\usage{
prepare_feature_instances(feature_instances_id, sample_sources, feature_atoms,
  output_dir)
}
\description{
Run by 'pymol instances.pml' in the distances directory.
}

