% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-save_tables.R
\name{save_tables}
\alias{save_tables}
\title{Save a data.frame as a table. For each output format,
 generate a table and put in the output directory}
\usage{
save_tables(features_analysis, table, table_id, sample_sources, output_dir,
  output_formats, table_title = NULL, caption = NULL, quote_strings = F,
  ...)
}
\description{
Save a data.frame as a table. For each output format,
 generate a table and put in the output directory
}

