% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-save_trees.R
\name{save_trees}
\alias{save_trees}
\title{Save a data.frame as a table. For each output format,
 generate a table and put in the output directory}
\usage{
save_trees(features_analysis, distance_matrix, tree_id, sample_sources,
  output_dir, output_formats, verbose = FALSE, ...)
}
\description{
Save a data.frame as a table. For each output format,
 generate a table and put in the output directory
}

