% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-density_estimation.R
\name{sliding_windows}
\alias{sliding_windows}
\title{Extract multiple, overlapping subsets of the data based on quantiles}
\usage{
sliding_windows(data, id.vars, measure.var, n_windows = 5,
  overlap_fraction = 0.3, verbose = FALSE)
}
\description{
Output: The rows of the input data duplicated once for each window
 it is contained in. The window id is in a column 'window.id' numbered
 1, 2, 3, ...
}
\details{
[--------------------------------]
[----------]
      [----------]
           [----------]
                 [----------]
                      [----------]
}

