% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-comparison_statistics.R
\name{smooth_kl_divergence}
\alias{smooth_kl_divergence}
\title{Here the inputs are probabilities over the sample space
assume evenly spaced partition of points}
\usage{
smooth_kl_divergence(x, ad, bd)
}
\description{
Here the inputs are probabilities over the sample space
assume evenly spaced partition of points
}

