% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-coordinate_normalizations.R
\name{spherical_normalization}
\alias{spherical_normalization}
\title{normalize according to spherical volume unit
r, radial distance and p, polar angle}
\usage{
spherical_normalization(r, p)
}
\description{
normalize according to spherical volume unit
r, radial distance and p, polar angle
}

