% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-vector_math.R
\name{vector_crossprod}
\alias{vector_crossprod}
\title{input:  Two arrays each of dimension c(n, 3)
output: An array of dimension c(n,3)  where the i'th row of the output is
       the vector cross product of the i'th rows of the input vectors}
\usage{
vector_crossprod(a, b)
}
\description{
input:  Two arrays each of dimension c(n, 3)
output: An array of dimension c(n,3)  where the i'th row of the output is
       the vector cross product of the i'th rows of the input vectors
}

