% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-vector_math.R
\name{vector_dihedral}
\alias{vector_dihedral}
\title{input:  4 arrays each of dimension c(n, 3)
output: An array of dimension c(n, 1) which is the dihedral angle about v2 -> v3
       between v1->v2 to v3->v4 using the right hand rule.}
\usage{
vector_dihedral(v1, v2, v3, v4)
}
\description{
input:  4 arrays each of dimension c(n, 3)
output: An array of dimension c(n, 1) which is the dihedral angle about v2 -> v3
       between v1->v2 to v3->v4 using the right hand rule.
}

