% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-vector_math.R
\name{vector_distance}
\alias{vector_distance}
\title{input: two arrays each of dimension c(n, d) each representing points in d dimensions
output: An aray of dimension c(n,1) which is the euclidean distance between the between the points}
\usage{
vector_distance(a, b)
}
\description{
input: two arrays each of dimension c(n, d) each representing points in d dimensions
output: An aray of dimension c(n,1) which is the euclidean distance between the between the points
}

