% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-vector_math.R
\name{vector_dotprod}
\alias{vector_dotprod}
\title{input:  Two arrays of dimension c(n, d)
output: An array of dimension c(n, 1) where the i'th element of the output is the
       the dot product of the i'th rows of the input}
\usage{
vector_dotprod(a, b)
}
\description{
input:  Two arrays of dimension c(n, d)
output: An array of dimension c(n, 1) where the i'th element of the output is the
       the dot product of the i'th rows of the input
}

