% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-vector_math.R
\name{vector_normalize}
\alias{vector_normalize}
\title{input:  An array of dimensions c(n, d)
output: An array of dimension c(n, d) where if the i'th row of the input is (x1,x2, ..., xd)
        then the i'th of the output is (x1/norm, x2/norm, ..., xd/norm) where
        norm = sqrt(x1^2 + x2^2 + ... + xd^2)}
\usage{
vector_normalize(m)
}
\description{
input:  An array of dimensions c(n, d)
output: An array of dimension c(n, d) where if the i'th row of the input is (x1,x2, ..., xd)
        then the i'th of the output is (x1/norm, x2/norm, ..., xd/norm) where
        norm = sqrt(x1^2 + x2^2 + ... + xd^2)
}

